CalcInterest - Question 1 Task 3

+++ CalcInterest/CalcInterest.java
@@ -1,56 +1,55 @@
-
 package CalcInterest;
 
 public class CalcInterest {
 
     // Method to compute loan interest
-    public static float computeLoanInterest(double loanAmount, int yearLoan, int loanType) {
+    public static double computeLoanInterest(double loanAmount, int yearLoan, int loanType) {
         if (loanAmount <= 0 || yearLoan <= 0 || (loanType != 1 && loanType != 2)) {
-            return -1; // Error case
+            return -1; // Error case for invalid input
         }
-        
-        float interestRate = 0;
-        
+
+        double interestRate = 0;
+
         if (loanType == 1) { // Home loan
             if (loanAmount < 100000) {
                 if (yearLoan <= 5) {
-                    interestRate = 8.0f;
+                    interestRate = 8.0;
                 } else if (yearLoan <= 10) {
-                    interestRate = 6.5f;
+                    interestRate = 6.5;
                 } else {
-                    interestRate = 5.5f;
+                    interestRate = 5.5;
                 }
             } else if (loanAmount < 500000) {
                 if (yearLoan <= 10) {
-                    interestRate = 6.5f;
+                    interestRate = 6.5;
                 } else {
-                    interestRate = 5.5f;
+                    interestRate = 5.5;
                 }
             } else {
-                interestRate = 5.5f;
+                interestRate = 5.5;
             }
         } else if (loanType == 2) { // Property loan
             if (loanAmount < 100000) {
                 if (yearLoan <= 5) {
-                    interestRate = 12.0f;
+                    interestRate = 12.0;
                 } else if (yearLoan <= 10) {
-                    interestRate = 8.5f;
+                    interestRate = 8.5;
                 } else {
-                    interestRate = 7.0f;
+                    interestRate = 7.0;
                 }
             } else if (loanAmount < 500000) {
                 if (yearLoan <= 10) {
-                    interestRate = 8.5f;
+                    interestRate = 8.5;
                 } else {
-                    interestRate = 7.0f;
+                    interestRate = 7.0;
                 }
             } else {
-                interestRate = 7.0f;
+                interestRate = 7.0;
             }
         }
 
         // Calculate total interest
-        float totalInterest = (float) (loanAmount * interestRate * yearLoan / 100);
+        double totalInterest = loanAmount * interestRate * yearLoan / 100;
         return totalInterest;
     }
